fix discount

+++ src/main/java/com/midorimart/managementsystem/entity/Product.java
@@ -58,7 +58,7 @@ public class Product {
     private Date created_at;
     private Date updated_at;
     private int deleted;
-    private double star;
+    private int star;
 
     @ManyToOne
     @JoinColumn(name = "country_code")
